[Exchange Oracle] Refactor crons structure
