+++ packages/examples/cvat/exchange-oracle/tests/integration/cron/state_trackers/test_track_completed_projects.py
@@ -4,7 +4,7 @@ import uuid
 from sqlalchemy.sql import select
 
 from src.core.types import Networks, ProjectStatuses, TaskStatuses, TaskTypes
-from src.crons.state_trackers import track_completed_projects
+from src.crons.cvat.state_trackers import track_completed_projects
 from src.db import SessionLocal
 from src.models.cvat import Project, Task
